requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2
django-filter==24.2
djangorestframework==3.15.1
importlib_metadata==7.1.0
Markdown==3.6
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
zipp==3.19.2
